requirements.txt
Django==4.0.2
djangorestframework==3.13.1
python-dotenv==0.19.2

